add loading indicators to admin pages and put some more data patterns in place (#18)
This adds loading indicators to the resource list page and the resource details page. It does this by utilizing SvelteKit's built-in streaming functionality. This lets us still do SSR but send Promises from the server to the client so the data can be fetched in the background while spinners show. There are still some rough edges with the streaming functionality, though. The main issue is that if the Promise rejects instead of resolving, the server will crash.

To work around this limitation and make working with the streamed data super easy, I've added some helper functions.

Additionally, this uses <a> in places where before we were using goto. The advantage with this is that SvelteKit will detect when the user hovers over a link and prefetch the data in the background, making the page transition feel a lot faster when the user clicks on the link.

Finally, this updates the resource list page to put the filters, page numbers, etc, into the search params in the URL. This helps with the <a> prefetching mentioned above. Some search param helper functions have also been added.