improve search param handling on resource list page (#29)
When adding the loading indicators and switching to using search params on the resource list page, I
missed the requirement that the page should reset to `1` when changing a filter or searching.

Attempting to fix this went down a rabbit hole of complexity due to the fact we need to debounce the
page URL changes or it would hit a race condition. Also Jesse pointed out that we want to avoid
`get` as much as possible when accessing stores. Because of all of this, I revisited a library I'd
looked at a couple weeks ago and was able to just copy/paste the code and make some needed changes
to make it work for our use case. It cleans things up quite a bit and makes interacting with the
search params much easier.