add step to auth initialization to check if refresh token is valid
When using `localstorage` mode for Auth0, it doesn't have an expiration time the same way that a
cookie would. Also, the `isAuthenticated()` will still return true if the token exists (even if it's
expired). Because of this, to ensure that the refresh token we have in localstorage is still valid,
we need to call `getTokenSilently` when initializing the auth. If it fails then we know the token
has expired and the user needs to log in again.

This also adds error handling to the resource edit page for if the token is expired. This shouldn't
be a common case now but it could still happen if the user loaded the application before expiration
but the token expired while they were using it.

We probably need to add some sort of pattern with maybe a global modal that would show a specific
error for if the session is expired, and ask the user to log back in.