fetch the current user at the root layout level (#34)
In order to allow this authorized endpoint to be accessed by the SSR SvelteKit path, some
refactoring of auth stuff needed to happen. This sets us up well for any future admin
endpoints that require authorization.

The approach here is:
1) Initialize auth from the client and set a 5-minute access token as a cookie
2) Keep this cookie current by setting it every minute
3) When a new page is server-side rendered, there are two possible paths:
  a) The auth token cookie exists, in which case we forward it as a Bearer token to the API and the
     data can be fetched by the server.
  b) The auth token cookie does not exist (e.g. the user loads the page fresh after inactivity), in
     which case the data can't be fetched by the server, so the client ends up making the requests
     itself.

This also updates the resource content status enum to match the new server values.