pull permissions for current user and skip authorized fetches if missing permissions (#42)
Right now when going to the resource edit page as a read-only user, the page errors out. This is because we're fetching the /admin/users endpoint to populate the publish/aquiferize modal dropdowns. In the case where the user is read-only, they don't have permissions to access that endpoint. There are a couple possible fixes, one of which is just to swallow the 403s in some cases. However, the best solution is to just not make the fetch if we know the user doesn't have permission.

This also has the added benefit of giving us access to the user's permissions from any page, both during SSR as well as on the client.