add assign user button to resources page and use permissions instead of roles (#43)
The resource page now shows an assign user button if the user is able to assign the resource. Check
BIB-233 for info on when the user is able to assign content.

It also switches to using permissions instead of roles for checking authorization. The pattern for
that is to use `data.currentUser.can(Permission.PublishContent)`.