fix heights of resource page panes so they fit content and fix null user message (#73)
The resource page has panes on the left that were set to grow in height and expand to take up the
full space. By using fit-content we can make them smaller when they should be.

Also there was a slight regression introduced recently where after the user has been away for long
enough to expire the access token, when they refresh the dashboard they would see an error message.
This fixes it.