separate draft and published resource into different editors
In order to prevent odd issues revolving around state and editor history (undo/redo), this separates
out the draft and published versions so they each live in a separate Tiptap editor.

Additionally it updates the buttons so that auto-save works when italics or bold or whatever is
used.

Finally, it fixes the undo and redo buttons so they're disabled when there is not history.