add back injected fetches to silence warnings and improve logging (#159)
For whatever reason, even though the injected fetch inside +page.ts files is only relevant to SSR, when you disable SSR you continue to get warnings about not using it. To fix this we'll just add it back in, since it's not hurting anything.

Also improves a couple logging areas so we aren't getting extra logs.