add sort back into URL search params (#166)
Reverts the change in a previous commit that removed the sort search param.

To handle the multiple tabs with different sort defaults, the change is made when clicking the tab directly. This let's us keep the sort param so that the back button preserves state, but also default the sort correctly per tab.