add content diffing between left and right versions (#182)
The underlying htmldiff-ts library is here:
https://github.com/BiblioNexusStudio/htmldiff-ts. I ported it from PHP
using AI and fixing mostly regex issues after the fact. I removed some
complicated list and table handling in the PHP code in favor of keeping
it simpler. It isn't the most performant, but since we're diffing in a
web worker it doesn't affect the main thread. I might look into
performance improvements once this merges and see if there are any easy
wins.