add function to fetch a bible passage (#202)
Some testing is showing anywhere from 250 - 650ms for the book of Psalms
for the first call. Then the cache kicks in and it's <1ms. Also, after
refresh the browser cache will have the data and the time is about 15ms.
So we're really only paying the data cost once per user per book per day
(which is the cache duration on the server and can probably be pushed
even higher).