addedFiltersAndColumns (#256)
@MrGrinst this refactor is for you 😉 

I did not like the use of the `canOnlyViewProjectsInCompany` variable.
Here's my reasoning:

1. It was hard to read. IE, sometimes it was negated, and sometimes it
wasn't. 🤷
2. It was being prop drilled.
3. It was duplicating work already done in the Auth Store. 
4. Handing the logic over to the Auth Store will make our codebase more
consistent.

This refactor makes sense because Amber will be on the project page
testing the new features. Knowing her, she will switch users to ensure
the manager and publisher can see what they should. This will not only
validate the new work but also the refactor.