replace text direction plugin with our own plugin (#278)
The existing text direction plugin had an issue where it always applied
the `dir` attribute to every node, which was redundant in most cases. It
included a `defaultDirection` option that could be set, but we weren't
using the option. Also the option only prevented the attribute from
getting applied to the HTML element. It was still in the underlying
JSON.

This caused two issues:
1. When a user went to edit a resource that hadn't had any changes yet,
the `dir` would be applied to the JSON and the editor thought a change
had been made. It would therefore send a PATCH update to the server,
even though no content had really changed.
2. When using AI translate on English to RTL languages, `dir="ltr"` was
being passed in the HTML and the RTL language was set in stone as LTR.

Additionally, the plugin needed a transaction to happen before it would
detect the LTR/RTL direction of nodes. This meant in order to generate
diff HTML (for instance), we had to do this hacky async function that
spun up a new editor, pretended to make a change so a transaction
happened, then waited a bit for the transaction to be applied before
grabbing the HTML.

By copying the plugin to an internal file we address all of these issues
and remove the need for the `dir` attribute to even be applied in the
majority of cases. The only times it will really even be set is on the
English source text of an RTL resource.