Bib 710 modify view edit page for community reviewer users (#427)
I modified
`src/routes/resources/[resourceContentId]/TranslationSelector.svelte` to
handle the value of the `languageid` as a number to take advantage of
Svelte's type inference and help us avoid calling `toString()` and
`parseInt()` when interacting with language Ids. This made sense as a
language id is only ever a string when used in the value of the `option`
html element and is never a string on the server.