Remove tap preload (#443)
The tap preload didn't differentiate between a regular click and a
middle-click / ctrl+click (i.e. a click to open a new tab). So any time
someone would do that, it would hit the API twice. This surely
contributes to the performance issues with the database when someone
quickly opens several items at once as new tabs.