switch to using aquifer-tiptap to add support for SIL Open Translators Notes (#471)
As we add more custom nodes/marks to our Tiptap setup, it makes more and
more sense for a single package to hold all of that configuration. The
chosen approach is for the `aquifer-tiptap` package to hold all of our
custom stuff (as well as lists of the nodes and marks we use) and then
our JSON/HTML utils, CMS, and Bible Well can all use a consistent
config.