simplify API calls and promise-based loading (#512)
Initially, due to using SSR, our API calls were mostly wrapped with a `{
promise: ... }` object using the `getFromApiWithoutBlocking` function.
This helped the pages feel like they loaded a bit more snappy when SSR
was turned on and we could have a spinner/skeleton on any part of the
page.

However, we were only really taking advantage of the spinner/skeleton in
two places, and we dropped SSR support a while back. Because of this, we
can simplify the API calls a lot and avoid a bunch of async/await code
by making the requests blocking again.

It's also worth pointing out that "blocking" in this case means that the
centered spinner shows up in the middle of the page while data is being
fetched. Previously with the SSR approach we would get a white page with
"blocking" loading, but now the HTML is there so we have a centered
spinner.

As a byproduct of doing more blocking calls in the `load` functions, the
error handling needed to be adjusted. It ended up being a simpler
approach and less boilerplatey, with a `errorGotoPath` global store that
can be set in a page's `load` function to indicate 404 redirect
behavior. Most other error handling is now done automatically through
the `+error.svelte` component.