Svelte 5 Upgrade (#532)
This pull request is heavily based on [this
guide](https://svelte.dev/docs/svelte/v5-migration-guide).

Running `npx sv migrate svelte-5` lead to a Svelte Kit upgrade that lead
to a few new linting errors caused by Svelte Kit 2's stricter linting.
And finally the migration script results.

## update
after running `npx sv migrate svelte-5` for the final upgrade of all
features, the application crashed in many places. It still functions but
obviously would not pass QA. I will try to manually take care of some of
the warnings that I see in the console.

## final update
After trying to manually resolve errors created by the script, I went on
to manually converting as many components to svelte 5. I ran into issues
with the following:

- unorthodox implementations of the svelte store. I believe the stores
will have to be completely refactored to match the documentation.
- Named slots are not backward compatible with svelte four, and any
component that calls an element with a named slot must also be changed.
This was a big issue for the Table and Modal components since larger
`+page.svelte` components use them.