Bib 844 for community reviewers only show helpful how to documents (#567)
Most of the Help content does not apply to Community Users. So, this
ticket has us implementing a `Help` tab on their Dashboard.

I used a `localstorage` flag to simulate a 'first login' check. If they
clear the storage in their browser, the disruption is minimal and this
prevents us from having to build out a backend solution for login
tracking.