Bib 834 showplay audio on aquifer resource pages (#566)
Should work like the ticket requires save for:
> The popup should close if the user clicks outside of it and the audio
should stop play.

To provide a consistent experience, I am utilizing the `Modal`
component. So, the 'click outside' would have to use an awkward
workaround of designating wrapper `div` just to track the bounds. Open
to suggestions that make sense and do not modify the `Modal`.

All FIA steps track play time elapsed. If a step is missing audio, the
player modal will inform the user of this. If the user device does not
support `webm` or `mp3`, the modal will inform the user of this.