Allow choosing both source and target languages when creating a project. (#631)
See also https://github.com/BiblioNexusStudio/aquifer-server/pull/737.

In addition to the requirements on BIB-906 I did a few extra things:
* Allow searching for project resource contents without specifying
chapters (defaults to "all" when you click search).
* Update code to allow for future Aquiferization of non-English resource
contents. It's still gated to English-only on both the front-end and
back-end but turning it on for other languages will be as simple as
removing one `filter()` call.
* When "Create project from already translated items" is true then don't
display English in the language dropdown (the back-end errors if you
pass English).