Bib 872 add at mentions in comments (#635)
`@mentions` can be seen in 3 places:

- Comment popup (resource page).
- Comments pane (resource page).
- Comment notifications (notifications tab on the dashboard)

Please note that when you insert a mention there will be a space
following the mention. If you try to remove the space, for example, to
place punctuation after the mention, it will read this as you intending
to edit the mention and not place the punctuation next to the mention.
 
Rather than spending more time on this nuanced issue, I've opted to get
this PR out there and get feedback.